autoconfCacheHook: init new hook
based on https://github.com/NixOS/nixpkgs/pull/172395 by @KAction
related: https://github.com/NixOS/nixpkgs/issues/172133

This change significantly improves the performance of the configurePhase for most autoconf based packages by using a pre-computed cache for its checks that is shared by all packages built via a specific stdenv.

`autoconf` is legacy infrastructure which was built around the idea that build environments should adapt to the current system. Nix' approach represents the opposite, providing a stable build environment independent of the current system. Much of autoconf is therefore, strictly speaking, unnecessary. Yet, we cannot easily get rid of it due to many projects depending on it.

`./configure` scripts generated by autoconf are simple shell scripts and don't benefit at all by multiple CPUs. Fortunately, these scripts support a `--cache-file` which can be used to inject results of a previous run of `./configure`, resulting in many checks being skipped on the second run.

Currently, we execute a large number of checks in each single derivation which uses autoconf. Many of these checks depend on inputs like the compiler or the libc which are stable across a stdenv. We can therefore cache many of these checks and only execute them once.

TL;DR; The cache needs to be curated.
Unfortunately it is not as straight forward as simply generating the cache. Some checks that autoconf performs, like, for example, checking for the existence of certain header files, should not be cached, as those can be modified by build inputs on each single derivation.
The curation is done by maintaining a global ignore list containing result variable names which are removed from the global cache before use.

- add cache file derivation that performs many tests to populate the cache. (nix-build -A autoconfCacheHook.cacheFile)
- add new hook `autoconfCacheHook` that injects the cache via configureFlags
- add the hook to the stdenv via stage 6 so it is used for all packages by default
- rebuild nixos many times and add entries to the ignore list which break builds

To debug bad cache entries I added debugging code to the hook's shell script.
By setting `autoconfCacheDiff = true` in a derivation, a diff between the result of `./configure` executed with vs. without the cache is printed to stdout.
This makes it easy to identify checks that have different results in bare stdenv vs a specific derivation. These checks can then be added to the global ignore list.

It ended up being quite a bit of shell code, but I separated the critical parts from the ones only being used for debugging. Only roughly the first 50 lines in `autoconf-cache-hook.sh` is considered critical code (see separator comment).

I am puzzled about why some of the results like `ac_cv_member_struct_utmpx_ut_name` are not stable across derivations and need to be ignored (see code comments in the ignore list). This specific check, for example, should only depend on glibc, which should't change across stdenv, yet its result does seem to change. This is not a big problem per se, but if there is actually an inconsistency in teh builds envs that can be prevented, or detected, doing so might be preferable from maintaining a larger ignore list.

I also would like someone to review how the hook is injected into stdenv. I am not entirely sure if it survives stdenv adapters/overrides.

There are potentially many more checks that can be safely added to the cache to exploit it even further. Though this is a low hanging fruit once the infrastructure is in place and can be added later.